+++ packages/vendure-plugin-webhook/src/api/webhook.resolver.ts
@@ -1,26 +1,27 @@
 import {
   Args,
   Mutation,
-  Query,
-  Resolver,
-  ResolveField,
   Parent,
+  Query,
+  ResolveField,
+  Resolver,
 } from '@nestjs/graphql';
 import {
   Allow,
   Ctx,
+  ForbiddenError,
+  Permission,
   PermissionDefinition,
-  ProductEvent,
   RequestContext,
 } from '@vendure/core';
-import { WebhookService } from './webhook.service';
 import {
   Webhook,
   WebhookInput,
   WebhookRequestTransformer,
 } from '../generated/graphql-types';
-import { Webhook as WebhookEntity } from './webhook.entity';
 import { RequestTransformer } from './request-transformer';
+import { Webhook as WebhookEntity } from './webhook.entity';
+import { WebhookService } from './webhook.service';
 
 // Permission needs to be defined first
 export const webhookPermission = new PermissionDefinition({
@@ -41,6 +42,30 @@ export class WebhookResolver {
     @Ctx() ctx: RequestContext,
     @Args('webhooks') webhooks: WebhookInput[]
   ): Promise<Webhook[]> {
+    const currentWebhooks = await this.webhookService.getAllWebhooks(ctx);
+    const newWebhooks = webhooks.filter(
+      (webhook) =>
+        !currentWebhooks.some(
+          (w) =>
+            w.channelAgnostic &&
+            w.event === webhook.event &&
+            w.url === webhook.url
+        )
+    );
+    // Check if any of the new webhooks are being set as channel-agnostic
+    const hasChannelAgnosticWebhook = newWebhooks.some(
+      (webhook) => webhook.channelAgnostic === true
+    );
+    if (hasChannelAgnosticWebhook) {
+      // Check if the current user has the "CreateChannel" permission
+      const channelPermissions = ctx.session?.user?.channelPermissions || [];
+      const hasPermission = channelPermissions.some((perm) =>
+        perm.permissions?.includes(Permission.CreateChannel)
+      );
+      if (!hasPermission) {
+        throw new ForbiddenError();
+      }
+    }
     return this.webhookService.saveWebhooks(ctx, webhooks);
   }